[Silver III] Title: 주유소, Time: 424 ms, Memory: 36284 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -9,15 +9,15 @@ public class Main  {
         StringTokenizer st = null;
 
         int n = Integer.parseInt(br.readLine()); // 도시의 개수
-        int[] far = new int[n-1]; // 도시 사이의 거라
-        int[] cost = new int[n]; // 각 도시의 기름값
-        int result = 0;
+        long[] far = new long[n-1]; // 도시 사이의 거라
+        long[] cost = new long[n]; // 각 도시의 기름값
+        long result = 0;
 
         st = new StringTokenizer(br.readLine());
         for(int i = 0; i<n-1; i++){
             far[i] = Integer.parseInt(st.nextToken());
         }
-        
+
         st = new StringTokenizer(br.readLine());
         for(int i = 0; i<n; i++){
             cost[i] = Integer.parseInt(st.nextToken());
